refactor: split auth hook and adjust server logic
