+++ .env.local
@@ -1,0 +1,2 @@
+DATABASE_URL=postgres://localhost/mydb
+NEXT_PUBLIC_SELECTED_USER_ID=00000000-0000-0000-0000-000000000001
