+++ app/api/admin/talent/trust_score/recalculate-all/route.ts
@@ -10,11 +10,14 @@ type SessionClaimsWithRole = {
 };
 
 export async function POST(_req: NextRequest) {
-  const { userId, sessionClaims } = await auth();
+  const isMock = process.env.NEXT_PUBLIC_USE_MOCK === 'true';
+  const { userId, sessionClaims } = isMock
+    ? { userId: 'mock', sessionClaims: { metadata: { role: 'admin' } } }
+    : await auth();
   const role = (sessionClaims as SessionClaimsWithRole)?.metadata?.role;
 
   // Check if user is authenticated and has admin role
-  if (!userId || role !== 'admin') {
+  if (!isMock && (!userId || role !== 'admin')) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
