+++ app/api/clients/[id]/projects/route.ts
@@ -8,11 +8,14 @@ type RouteContext = { params: Promise<{ id: string }> };
 
 export async function GET(_req: NextRequest, ctx: RouteContext) {
   const { id } = await ctx.params;
-  const { userId, sessionClaims } = await auth();
+  const isMock = process.env.NEXT_PUBLIC_USE_MOCK === 'true';
+  const { userId, sessionClaims } = isMock
+    ? { userId: 'mock', sessionClaims: { metadata: { role: 'admin' } } }
+    : await auth();
   const role = (sessionClaims as SessionClaimsWithRole)?.metadata?.role;
   
   // Check if user is authenticated
-  if (!userId) {
+  if (!isMock && !userId) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
   
@@ -20,7 +23,7 @@ export async function GET(_req: NextRequest, ctx: RouteContext) {
   const isAdmin = role === 'admin';
   const isOwnProfile = userId === id;
   
-  if (!isAdmin && !isOwnProfile) {
+  if (!isMock && !isAdmin && !isOwnProfile) {
     return NextResponse.json({ error: 'Forbidden' }, { status: 403 });
   }
