+++ app/api/clients/[id]/trust-score/route.ts
@@ -2,7 +2,7 @@ import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
 import { db } from '@/lib/db';
 import { users } from '@/lib/schema';
-import { eq } from 'drizzle-orm/pg-core';
+import { eq } from 'drizzle-orm';
 import { auth } from '@clerk/nextjs/server';
 import type { SessionClaimsWithRole } from '@/lib/types';
 
@@ -13,11 +13,14 @@ export async function POST(
   ctx: RouteContext
 ) {
   const { id } = await ctx.params;
-  const { userId, sessionClaims } = await auth();
+  const isMock = process.env.NEXT_PUBLIC_USE_MOCK === 'true';
+  const { userId, sessionClaims } = isMock
+    ? { userId: 'mock', sessionClaims: { metadata: { role: 'admin' } } }
+    : await auth();
   const role = (sessionClaims as SessionClaimsWithRole)?.metadata?.role;
   
   // Check if user is authenticated and has admin role
-  if (!userId || role !== 'admin') {
+  if (!isMock && (!userId || role !== 'admin')) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
