+++ app/api/db/route.ts
@@ -20,15 +20,16 @@ const tableMap = {
   escrow_transactions: escrowTransactions,
   notifications,
 } as const;
-import { eq } from 'drizzle-orm/pg-core';
+import { eq } from 'drizzle-orm';
 import { auth } from '@clerk/nextjs/server';
 import type { SessionClaimsWithRole } from '@/lib/types';
 
 export async function GET(request: NextRequest) {
-  const { userId } = await auth();
+  const isMock = process.env.NEXT_PUBLIC_USE_MOCK === 'true';
+  const { userId } = isMock ? { userId: 'mock' } : await auth();
   
   // Check if user is authenticated
-  if (!userId) {
+  if (!isMock && !userId) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
   
@@ -61,10 +62,11 @@ export async function GET(request: NextRequest) {
 }
 
 export async function POST(request: NextRequest) {
-  const { userId } = await auth();
+  const isMock = process.env.NEXT_PUBLIC_USE_MOCK === 'true';
+  const { userId } = isMock ? { userId: 'mock' } : await auth();
   
   // Check if user is authenticated
-  if (!userId) {
+  if (!isMock && !userId) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
   
@@ -90,10 +92,11 @@ export async function POST(request: NextRequest) {
 }
 
 export async function PUT(request: NextRequest) {
-  const { userId } = await auth();
+  const isMock = process.env.NEXT_PUBLIC_USE_MOCK === 'true';
+  const { userId } = isMock ? { userId: 'mock' } : await auth();
   
   // Check if user is authenticated
-  if (!userId) {
+  if (!isMock && !userId) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
   
@@ -119,11 +122,14 @@ export async function PUT(request: NextRequest) {
 }
 
 export async function DELETE(request: NextRequest) {
-  const { userId, sessionClaims } = await auth();
+  const isMock = process.env.NEXT_PUBLIC_USE_MOCK === 'true';
+  const { userId, sessionClaims } = isMock
+    ? { userId: 'mock', sessionClaims: { metadata: { role: 'admin' } } }
+    : await auth();
   const role = (sessionClaims as SessionClaimsWithRole)?.metadata?.role;
   
   // Check if user is authenticated and has admin role
-  if (!userId || role !== 'admin') {
+  if (!isMock && (!userId || role !== 'admin')) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
