+++ app/api/projects/route.ts
@@ -3,12 +3,13 @@ import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
 
 export async function GET(_req: NextRequest) {
-  const { userId, sessionClaims } = await auth();
+  const isMock = process.env.NEXT_PUBLIC_USE_MOCK === 'true';
+  const { userId, sessionClaims } = isMock ? { userId: 'mock', sessionClaims: { metadata: { role: 'admin' } } } : await auth();
 
   // Cast to ensure TypeScript knows the shape of sessionClaims.metadata
   const role = (sessionClaims?.metadata as { role?: string })?.role;
 
-  if (!userId || role !== 'admin') {
+  if (!isMock && (!userId || role !== 'admin')) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
