+++ app/talent/dashboard/page.tsx
@@ -6,7 +6,7 @@ import { Card, CardContent } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Badge } from "@/components/ui/badge";
 import { toast } from "sonner";
-import { useAuth } from "@/lib/useAuth";
+import { useAuth } from "@/lib/client/useAuthContext";
 import { useMockData } from "@/lib/useMockData";
 import {
   Clock,
