+++ lib/apiHandlers/talent.ts
@@ -1,5 +1,5 @@
 import { db } from '@/lib/db';
-import { eq } from 'drizzle-orm/pg-core';
+import { eq } from 'drizzle-orm';
 import { talentProfiles } from '@/lib/schema';
 
 export async function getTalentById(id: string) {
