+++ lib/db/users.ts
@@ -1,6 +1,6 @@
 import { db } from '../db';
 import { users } from '../schema';
-import { eq } from 'drizzle-orm/pg-core';
+import { eq } from 'drizzle-orm';
 
 export type User = typeof users.$inferSelect;
