+++ server/selectUserByRole.ts
@@ -1,6 +1,6 @@
 import { db } from '../lib/db';
 import { users } from '../lib/schema';
-import { eq } from 'drizzle-orm/pg-core';
+import { eq } from 'drizzle-orm';
 
 export type TestRole = 'admin' | 'client' | 'talent';
